change ready condition logic
