+++ images/virtualization-artifact/pkg/controller/vd/vd_controller.go
@@ -87,7 +87,6 @@ func NewController(
 		internal.NewAttacheeHandler(mgr.GetClient()),
 		internal.NewStatsHandler(stat, importer, uploader),
 		internal.NewInUseHandler(mgr.GetClient()),
-		internal.NewQuotaExceededHandler(mgr.GetClient()),
 	)
 
 	vdController, err := controller.New(ControllerName, mgr, controller.Options{
